# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Linha",
    "version": "0.1.0",
    "description": "Versioned JSON base protocol. SDKs wrap these operations. Result templates and policies are snapshotted; job IDs survive process restarts. Authentication follows deployment configuration: security.enabled=false disables client and worker authentication; security.oidc.enabled=false disables only client authentication. Anonymous clients share owner anonymous. Secured deployments require the bearer schemes below."
  },
  "paths": {
    "/v1/contexts:ensure": {
      "post": {
        "operationId": "ensureContext",
        "summary": "Persist an immutable owned backend specification; readiness is separate",
        "security": [
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Context"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnsureRequest"
              },
              "example": {
                "name": "metoc-v2",
                "spec": {
                  "image": "registry.example/metoc/worker:1",
                  "engine": {
                    "type": "spark",
                    "version": "3.5.6",
                    "settings": {
                      "kubernetes": {
                        "driverPodTemplate": {
                          "spec": {
                            "imagePullSecrets": [
                              {
                                "name": "metoc-registry"
                              }
                            ],
                            "volumes": [
                              {
                                "name": "app-config",
                                "configMap": {
                                  "name": "metoc-config-v1"
                                }
                              }
                            ],
                            "containers": [
                              {
                                "name": "main",
                                "env": [
                                  {
                                    "name": "APP_ROLE",
                                    "value": "driver"
                                  }
                                ],
                                "volumeMounts": [
                                  {
                                    "name": "app-config",
                                    "mountPath": "/etc/metoc",
                                    "readOnly": true
                                  }
                                ]
                              }
                            ]
                          }
                        },
                        "executorPodTemplate": {
                          "spec": {
                            "imagePullSecrets": [
                              {
                                "name": "metoc-registry"
                              }
                            ],
                            "volumes": [
                              {
                                "name": "app-config",
                                "configMap": {
                                  "name": "metoc-config-v1"
                                }
                              }
                            ],
                            "containers": [
                              {
                                "name": "main",
                                "env": [
                                  {
                                    "name": "APP_ROLE",
                                    "value": "executor"
                                  }
                                ],
                                "volumeMounts": [
                                  {
                                    "name": "app-config",
                                    "mountPath": "/etc/metoc",
                                    "readOnly": true
                                  }
                                ]
                              }
                            ]
                          }
                        }
                      }
                    }
                  },
                  "results": {
                    "type": "local",
                    "root": "/var/lib/linha/results"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/contexts/{context}": {
      "get": {
        "operationId": "getContext",
        "summary": "getContext",
        "security": [
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Context"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "context",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/contexts/{context}/jobs": {
      "post": {
        "operationId": "submitJob",
        "summary": "Return an ID after durable acceptance, before engine execution",
        "security": [
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "context",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Linha-Client-Lease",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Live lease returned by ensure or attach. Idempotent replay is scoped by owner/logical name/key across versions."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          }
        }
      }
    },
    "/v1/jobs": {
      "get": {
        "operationId": "listJobs",
        "summary": "listJobs",
        "security": [
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "contextId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ]
      }
    },
    "/v1/jobs/{job}": {
      "get": {
        "operationId": "getJob",
        "summary": "getJob",
        "security": [
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/v1/jobs/{job}/result": {
      "get": {
        "operationId": "getResult",
        "summary": "Fetch a retained result descriptor; pending results return RESULT_NOT_READY",
        "security": [
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/jobs/{job}/files/{output}": {
      "get": {
        "operationId": "downloadResultFile",
        "summary": "downloadResultFile",
        "security": [
          {
            "clientBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized streamed file bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "job",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "output",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/v1/contexts/{context}/dev-workers": {
      "post": {
        "operationId": "createLocalWorkerIdentity",
        "summary": "Security disabled only: allocate local worker routing identity (empty token)",
        "security": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Structured error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "context",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "JobPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
...
    },
    "securitySchemes": {
      "clientBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated OIDC token; explicitly configured development token only in development mode"
      },
      "workerBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Credential scoped to worker context and incarnation"
      },
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "linha_session",
        "description": "Opaque HA browser session; viewer/operator mapping checked for each admin request. Cookie mutations additionally require Origin and X-Linha-CSRF."
      }
    }
  }
}